# openapi/v1.json
{
  "components": {
    "schemas": {
...
      "ApiErrorResponseDto": {
        "properties": {
          "code": {
            "type": "string"
          },
          "correlationId": {
            "format": "uuid",
            "type": "string"
          },
          "fieldErrors": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "retryAfterSeconds": {
            "minimum": 1,
            "type": "number"
          },
          "retryable": {
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable",
          "correlationId"
        ],
        "type": "object"
      },
...
      "CustomerResolvedPriceResponseDto": {
        "properties": {
          "amountMinor": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "source": {
            "enum": [
              "customer_specific",
              "global"
            ],
            "type": "string"
          },
          "status": {
            "enum": [
              "resolved",
              "missing"
            ],
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
...
      "ResolvedPriceResponseDto": {
        "properties": {
          "amountMinor": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "source": {
            "enum": [
              "customer_specific",
              "global"
            ],
            "type": "string"
          },
          "sourcePriceId": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "enum": [
              "resolved",
              "missing"
            ],
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "opaqueBearer": {
        "bearerFormat": "opaque",
        "scheme": "bearer",
        "type": "http"
      },
      "refreshCookie": {
        "in": "cookie",
        "name": "milktrack_refresh",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "",
    "title": "MilkTrack API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/customer/vendors/{vendorId}/households/{householdId}/prices/resolved": {
      "get": {
        "operationId": "CustomerResolvedPriceController_resolve",
        "parameters": [
          {
            "in": "path",
            "name": "vendorId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "productId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "unitId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "deliverySlotId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "serviceDate",
            "required": true,
            "schema": {
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerResolvedPriceResponseDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "opaqueBearer": []
          }
        ],
        "tags": [
          "Customer resolved pricing"
        ]
      }
    },
...
    "/v1/vendors/{vendorId}/prices/resolved": {
      "get": {
        "operationId": "VendorResolvedPriceController_resolve",
        "parameters": [
          {
            "in": "path",
            "name": "vendorId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "productId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "unitId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "deliverySlotId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "serviceDate",
            "required": true,
            "schema": {
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolvedPriceResponseDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "opaqueBearer": []
          }
        ],
        "tags": [
          "Resolved pricing"
        ]
      }
    },
...
  },
  "servers": [],
...
}